+++ R/apriori.R
@@ -141,6 +141,9 @@
 #' 	parameter = list(supp = 0.5, conf = 0.9, target = "rules"))
 #' summary(rules)
 #'
+#' # since ... gets automatically added to parameter, we can also write the 
+#' #  same call shorter:
+#' apriori(Adult, supp = 0.5, conf = 0.9, target = "rules")
 #' @export apriori
 apriori <-
   function(data,